Create EnumerateRoleOperations.ps1
This script enumerates the role operations for a specific role. For example if you have something like Microsoft.Network/* or Microsoft.Network/VirtualNetwork/* this enumerates every operation till the last sub-tree. This is helpful in listing out all the available role operations and also in comparing two different roles if needed.